+++ app/src/main/java/com/example/hangoutz/utils/Extension.kt
@@ -2,11 +2,13 @@ package com.example.hangoutz.utils
 
 import android.annotation.SuppressLint
 import android.os.Build
+import android.util.Log
 import androidx.annotation.RequiresApi
 import com.example.hangoutz.data.models.Friend
 import java.text.SimpleDateFormat
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter.ofPattern
+import java.util.Date
 import java.util.Locale
 
 
@@ -70,4 +72,18 @@ fun convertTimeToMillis(dateTimeString: String? = null): Long {
     } else {
         System.currentTimeMillis()
     }
+}
+
+ fun checkIfInPast(date: String): Boolean {
+    var isValid: Boolean = false
+    val inputFormat = SimpleDateFormat("dd.MM.yyyy HH:mm", Locale.getDefault())
+    try {
+        val inputDateTime = inputFormat.parse(date)
+        if (inputDateTime != null && inputDateTime.before(Date())) {
+            isValid = true
+        } else isValid = false
+    } catch (e: Exception) {
+        Log.e("Error", "Exception while parsing")
+    }
+    return isValid
 }
